Create the connection program
Without a special exec-like program, each application would have to
make connection by itself.
That would be disastrous because it prevents rforking processes.

This will establish that connection the first time.

A fork program would be useful for testing the queuing algorithm.
A test program (cat) would be useful too.